benchmark: round-3 re-timing (Schur + fused extraction) — wp_ultra, README WP figure, MDBM type-assert
Re-measures the CPU/memory figures affected by the round-3 solver work (real-Schur
default + fused static coefficient extraction). All ρ/Var values are bit-identical;
only timings and memory change.

- benchmark_wp_ultra.jl: add WP_MF_ONLY mode (reuse the untouched classical and
  step-recursion rows, re-measure only the multiplication-free column) + min-of-N
  timing. At p=192 MF is now 0.030 s / 22 MB (Schur halves the Krylov-basis memory),
  so the explicit-product advantage grows to 1.3e4 in time and 1.8e4 in allocation.
- wp_ultra.csv: refreshed MF column.
- assets/HighOrderConvergence.png: regenerated (error vs CPU time, round-3 GL/SD).
- README.md: MF-vs-explicit-product ratio 8.5e3 -> 1.3e4.
- ssv2dof_chart.jl: assert ::Float64 on the scalar MDBM curve functions. The factored
  spectral-radius solver's return_vec keyword leaves its return type-inferred as a
  Union (boxed scalar), which broke MDBM boundary tracing (zero(::Any)); the runtime
  value is always Float64, so this is a value no-op. Root-cause type-stability fix
  tracked separately.

+++ benchmark/ssv2dof_chart.jl
@@ -120,10 +120,15 @@ if mode in ("bf","both")
     @printf("2DOF BF TOTAL: %.1f s\n", time()-t0); flush(stdout)
 end
 
+# The `::Float64` assertions pin the curve-function value type for MDBM: the map
+# builds its solution containers from the inferred return type, and the factored
+# spectral-radius solver's `return_vec` keyword makes its return type-inferred as
+# a `Union` (a boxed scalar), which would otherwise leave MDBM with `zero(::Any)`.
+# The runtime value is always a `Float64`, so the assertion is a no-op on values.
 CURVEDEFS = Dict{String,Function}(
-    "cs_det"  => (ξ,w) -> solve_point(10.0^ξ, w; rva=0.0, σcv=0.0, σav=0.0)[1] - 1.0,
-    "ssv_det" => (ξ,w) -> solve_point(10.0^ξ, w; σcv=0.0, σav=0.0)[1] - 1.0,
-    "ssv_sto" => (ξ,w) -> solve_point(10.0^ξ, w)[1] - 1.0,
+    "cs_det"  => (ξ,w) -> (solve_point(10.0^ξ, w; rva=0.0, σcv=0.0, σav=0.0)[1] - 1.0)::Float64,
+    "ssv_det" => (ξ,w) -> (solve_point(10.0^ξ, w; σcv=0.0, σav=0.0)[1] - 1.0)::Float64,
+    "ssv_sto" => (ξ,w) -> (solve_point(10.0^ξ, w)[1] - 1.0)::Float64,
     "ssv_var" => (ξ,w) -> begin
         ρ,v = solve_point(10.0^ξ, w; want_var=true)
         (ρ ≥ 0.999 || isnan(v)) ? 1.0 : (v - VARLIM)
